Fix issue with current state callback
